Updated UserToken and ApiOboToken services
