+++ src/DfE.CoreLibs.Security/Interfaces/IUserTokenService.cs
@@ -5,16 +5,8 @@ namespace DfE.CoreLibs.Security.Interfaces
     /// <summary>
     /// Interface for acquiring API access tokens.
     /// </summary>
-    public interface ITokenService
+    public interface IUserTokenService
     {
-        /// <summary>
-        /// Acquires an API access token by mapping the user's roles to API scopes/permissions using the On-Behalf-Of (OBO) flow.
-        /// </summary>
-        /// <param name="authenticationScheme">The authentication scheme.</param>
-        /// <returns>A JWT access token string for accessing the API.</returns>
-        /// <exception cref="UnauthorizedAccessException">Thrown if the user does not have any roles or if no valid scopes are found for the user's roles.</exception>
-        Task<string> GetApiOboTokenAsync(string? authenticationScheme = null);
-
         /// <summary>
         /// ----
         /// DO NOT USE THIS METHOD IF YOUR API IS RUNNING IN A SEPARATE SERVICE.
@@ -29,6 +21,5 @@ namespace DfE.CoreLibs.Security.Interfaces
         /// <exception cref="InvalidOperationException">Thrown when the user does not have a valid identifier.</exception>
         /// <exception cref="ArgumentNullException">Thrown when the <paramref name="user"/> is null.</exception>
         Task<string> GetUserTokenAsync(ClaimsPrincipal user);
-
     }
 }
